Login done

+++ app/Http/Controllers/CountryController.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 namespace App\Http\Controllers;
 
@@ -13,11 +13,11 @@ class CountryController
      */
     public function index(Request $request)
     {
-        $query  = Country::query();
+        $query = Country::query();
         if (!empty($request->search)) {
-            $query->where('name', 'LIKE', '%'. $request->search .'%')->orWhere('short_name', 'LIKE', '%'. $request->search .'%');
+            $query->where('name', 'LIKE', '%' . $request->search . '%')->orWhere('short_name', 'LIKE', '%' . $request->search . '%');
         }
-        $countries = Country::latest()->paginate(10);
+        $countries = $query->latest()->paginate(10);
         return view('admin.screens.country.index', compact('countries'));
     }
 
@@ -71,9 +71,9 @@ class CountryController
         $country->name = $request->name;
         $country->short_name = $request->short_name;
         $country->code = $request->code;
-        if(!empty($request->hasFile('flag'))){
-            if(!empty($country->flag)){
-                unlink(public_path() ."/storage/". $country->getRawOriginal('flag'));
+        if (!empty($request->hasFile('flag'))) {
+            if (!empty($country->flag) && file_exists(public_path() . "/storage/" . $country->getRawOriginal('flag'))) {
+                unlink(public_path() . "/storage/" . $country->getRawOriginal('flag'));
             }
             $country->flag = $request->flag->store("countries", "public");
         }
@@ -87,12 +87,26 @@ class CountryController
      */
     public function destroy(Country $country)
     {
-        if(!empty($country->flag)){
+        if (!empty($country->flag)) {
             // Remove from the folder
-            unlink(public_path() ."/storage/". $country->getRawOriginal('flag'));
+            unlink(public_path() . "/storage/" . $country->getRawOriginal('flag'));
         }
         $country->delete();
 
         return redirect()->back()->with('success', 'Success! A record has been deleted.');
+    }
+
+    public function bulkDelete(Request $request)
+    {
+        foreach ($request->delIds as $id) {
+            $country = Country::find($id);
+            $file = public_path() . "/storage/" . $country->getRawOriginal('flag');
+            if (!empty($country->flag) && file_exists($file)) {
+                unlink($file);
+            }
+            $country->delete();
+        }
+
+        return redirect()->back()->with('success', 'Success! Selected record(s) has been deleted.');
     }
 }
